コンポーネントを分けました
