+++ frontend/src/pages/ContestEditPage/index.tsx
@@ -3,17 +3,10 @@ import {
   FormLabel,
   SimpleGrid,
   Box,
-  Switch,
   FormControl,
-  Popover,
-  PopoverTrigger,
-  PopoverContent,
   Textarea,
   NumberInput,
   NumberInputField,
-  NumberInputStepper,
-  NumberIncrementStepper,
-  NumberDecrementStepper,
 } from '@chakra-ui/react';
 import PropTypes from 'prop-types';
 import { FC, useEffect, useState } from 'react';
@@ -23,7 +16,6 @@ import {
   Draggable,
   OnDragEndResponder,
 } from 'react-beautiful-dnd';
-import { MutableListElement } from '../../components/MutableListElement';
 import {
   getContestInfo,
   getContestProblems,
@@ -34,31 +26,13 @@ import {
 import { findContestIdFromPath } from '../../functions/findContestIdFromPath';
 import { getCookie } from '../../functions/getCookie';
 import { isMobile } from '../../functions/isMobile';
-import { ContestCreator, ProblemInfo } from '../../types';
+import { ContestCreator, ProblemInfo, EditProblemInfo } from '../../types';
 import './ContestEditPage.css';
+import ProblemEditColumnMobile from './ProblemEditColumnMobile';
+import ProblemEditColumnPC from './ProblemEditColumnPC';
 
 // URL: /contest/$contestName/edit
 
-class EditProblemInfo {
-  statement: string;
-  id: number;
-  point: number | undefined;
-  answer: string[];
-  isQuiz: boolean;
-  constructor(
-    statement: string,
-    id: number,
-    point: number,
-    answer: string[],
-    izQuiz: boolean
-  ) {
-    this.statement = statement;
-    this.id = id;
-    this.point = point;
-    this.answer = answer;
-    this.isQuiz = izQuiz;
-  }
-}
 interface EditProblemsElementProps {
   problems: EditProblemInfo[];
   setProblems(problems: EditProblemInfo[]): void;
@@ -69,192 +43,29 @@ const EditProblemsElement: FC<EditProblemsElementProps> = ({
   setProblems,
 }) => {
   const columnHeight = isMobile() ? 150 : 120;
-  const updateProblemStatement = (idx: number, statement: string) => {
-    const newProblems = [...problems];
-    newProblems[idx].statement = statement;
-    setProblems(newProblems);
-  };
-  const updateProblemQuizMode = (idx: number, isQuiz: boolean) => {
-    const newProblems = [...problems];
-    newProblems[idx].isQuiz = isQuiz;
-    setProblems(newProblems);
-  };
-  const updateProblemPoint = (idx: number, point: string) => {
-    const newProblems = [...problems];
-    let newPoint: number | undefined = parseInt(point);
-    if (isNaN(newPoint)) {
-      newPoint = undefined;
-    }
-    newProblems[idx].point = newPoint;
-    setProblems(newProblems);
-  };
-  const eraseProblem = (idx: number) => {
-    if (problems.length === 1) {
-      return;
-    }
-    const newProblems = problems.filter((_, problemIdx) => problemIdx !== idx);
-    setProblems(newProblems);
-  };
-  const addProblem = () => {
-    const newProblems = [...problems];
-    const newId = problems.slice(-1)[0].id + 1;
-    newProblems.push(new EditProblemInfo('', newId, 1, [''], false));
-    setProblems(newProblems);
-  };
-  const setNewAnswer = (idx: number, newAnswer: string[]) => {
-    const newProblems = [...problems];
-    newProblems[idx].answer = newAnswer;
-    setProblems(newProblems);
-  };
-  const UP_REARRANGE = 'UP';
-  const DOWN_REARRANGE = 'DOWN';
-  const rearrangeProblem = (idx: number, direction: string) => {
-    const newProblems = [...problems];
-    if (direction === UP_REARRANGE && idx !== 0) {
-      const tmp = newProblems[idx];
-      newProblems[idx] = newProblems[idx - 1];
-      newProblems[idx - 1] = tmp;
-    }
-    if (direction === DOWN_REARRANGE && idx !== newProblems.length - 1) {
-      const tmp = newProblems[idx];
-      newProblems[idx] = newProblems[idx + 1];
-      newProblems[idx + 1] = tmp;
-    }
-    setProblems(newProblems);
-  };
 
   const listGroups = problems.map((problem, idx) => {
-    return (
-      <SimpleGrid columns={5} spacing={10} key={problem.id}>
-        <Box>
-          <FormLabel>問題文</FormLabel>
-          <div className={'mb-3'}>
-            <Textarea
-              placeholder={'〇〇な△△な〜んだ？'}
-              value={problem.statement}
-              onChange={(e) => updateProblemStatement(idx, e.target.value)}
-            />
-          </div>
-        </Box>
-        <Box>
-          <div
-            style={{
-              height: columnHeight,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-            }}
-          >
-            <Popover>
-              <PopoverTrigger>
-                <Button colorScheme={'blue'}>答え編集</Button>
-              </PopoverTrigger>
-              <PopoverContent>
-                <div id={'popover-basic'}>
-                  <MutableListElement
-                    items={problem.answer}
-                    setItems={(newAnswer: string[]) => {
-                      setNewAnswer(idx, newAnswer);
-                    }}
-                  />
-                </div>
-              </PopoverContent>
-            </Popover>
-          </div>
-        </Box>
-        <Box>
-          <div
-            style={{
-              height: columnHeight,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-            }}
-          >
-            <div className={'mb-3'}>
-              <NumberInput
-                value={problem.point}
-                onChange={(value) => {
-                  updateProblemPoint(idx, value);
-                }}
-              >
-                <NumberInputField />
-                <NumberInputStepper>
-                  <NumberIncrementStepper
-                    onClick={() =>
-                      updateProblemPoint(
-                        idx,
-                        ((problem?.point || 0) + 1).toString()
-                      )
-                    }
-                  />
-                  <NumberDecrementStepper
-                    onClick={() =>
-                      updateProblemPoint(
-                        idx,
-                        ((problem?.point || 0) - 1).toString()
-                      )
-                    }
-                  />
-                </NumberInputStepper>
-              </NumberInput>
-            </div>
-            <p style={{ marginLeft: 12 }}>点</p>
-          </div>
-        </Box>
-        <Box>
-          {' '}
-          <div
-            style={{
-              height: columnHeight,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              textAlign: 'center',
-              fontSize: '1.125rem',
-            }}
-          >
-            <Switch
-              id={`${problem.id} switch`}
-              defaultChecked={problem.isQuiz}
-              onChange={(e) => {
-                updateProblemQuizMode(idx, e.target.checked);
-              }}
-            />
-            <FormLabel>Enable Quiz Mode</FormLabel>
-          </div>
-        </Box>
-        <Box>
-          <div
-            style={{
-              height: columnHeight,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-            }}
-          >
-            <Button type={'button'} onClick={addProblem}>
-              +
-            </Button>
-            <Button type={'button'} onClick={() => eraseProblem(idx)}>
-              -
-            </Button>
-            <Button
-              type={'button'}
-              onClick={() => rearrangeProblem(idx, UP_REARRANGE)}
-            >
-              ↑
-            </Button>
-            <Button
-              type={'button'}
-              onClick={() => rearrangeProblem(idx, DOWN_REARRANGE)}
-            >
-              ↓
-            </Button>
-          </div>
-        </Box>
-      </SimpleGrid>
-    );
+    if (isMobile()) {
+      return (
+        <ProblemEditColumnMobile
+          key={problem.id}
+          idx={idx}
+          setProblems={setProblems}
+          problems={problems}
+          height={columnHeight}
+        />
+      );
+    } else {
+      return (
+        <ProblemEditColumnPC
+          key={problem.id}
+          idx={idx}
+          setProblems={setProblems}
+          problems={problems}
+          height={columnHeight}
+        />
+      );
+    }
   });
 
   const onDragEnd: OnDragEndResponder = (result) => {
